feat: enhance error handling and user association in customer, order, and product services and repositories

+++ src/repositories/product.repository.ts
@@ -4,23 +4,25 @@ import { IProductRepository } from '../interfaces/IProductRepository'
 const prisma = new PrismaClient()
 
 export class ProductRepository implements IProductRepository {
-  async findAll() {
-    return prisma.products.findMany()
+  async findAll(user_id: string) {
+    return prisma.products.findMany({ where: { user_id } })
   }
   
-  async findById(id: number) {
-    return prisma.products.findUnique({ where: { id } })
+  async findById(id: number, user_id: string) {
+    return prisma.products.findUnique({ where: { id, user_id } })
   }
 
-  async create(data: any) {
+  async create(data: any, user_id: string) {
+    data.user_id = user_id
     return prisma.products.create({ data })
   }
 
-  async update(id: number, data: any) {
-    return prisma.products.update({ where: { id }, data })
+  async update(id: number, user_id: string, data: any) {
+    data.user_id = user_id
+    return prisma.products.update({ where: { id, user_id }, data })
   }
 
-  async delete(id: number) {
-    await prisma.products.delete({ where: { id } })
+  async delete(id: number, user_id: string) {
+    await prisma.products.delete({ where: { id, user_id } })
   }
 }
